Classy work around.
So now tests works.. The reason this works is because flask routing
passes the attribute using kwargs. In this case, the self parameter gets
passed into `args`. When we pass project, we pass via project=project
and the self gets passed into the correct location as a part of `*args`.

Probably not a perma-solution. Still gotta refer to the comments in #16.