Omg I found a ghost!
So it turns out, you're not suppose to inject `$route` with
`["$route", function($route)]` as it will load the default template or
something stupid like that...

This means that during unittests it will cause you a week of headache as
you have no fucking idea why there's an unexpected GET request in your
tests. Until you realize that `$route` is the cause.

The fix for this problem going forward is to either mock out `$route`
or not use it, as it doesn't seem to be useful for `TodosController`
anyway.

Anyways. I found the ghost that has been haunting me for a week. I think
I'm gonna go cry for a bit.